requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
dlib-bin==19.24.6
watchdog==3.0.0
